requirements.txt
PyYAML
colorlog
mido
